fix modal styling with dialog component
