+++ src/frontend/ExtensionManager/index.module.scss
@@ -11,19 +11,5 @@
   top: var(--top-navbar-height);
   background-color: transparent;
   border: none;
-  > div.mmWindowContainer {
-    top: var(--top-navbar-height);
-    position: fixed;
-    right: 0px;
-  }
-}
-
-.mmWindow {
-  width: 357px;
-  height: 600px;
-  overflow: hidden;
-  border-radius: 0px 0px 16px 16px;
-  background: #24272a;
-  display: flex;
-  box-shadow: rgb(0 0 0 / 35%) 0px 5px 15px;
+  outline: none;
 }
